cli/src/main/kotlin/org/coner/trailer/cli/command/policy/PolicyDeleteCommand.kt
package org.coner.trailer.cli.command.policy

import com.github.ajalt.clikt.core.CliktCommand
import com.github.ajalt.clikt.core.findOrSetObject
import com.github.ajalt.clikt.parameters.arguments.argument
import com.github.ajalt.clikt.parameters.arguments.convert
import org.coner.trailer.cli.util.clikt.toUuid
import org.coner.trailer.io.service.PolicyService
import org.kodein.di.DI
import org.kodein.di.DIAware
import org.kodein.di.instance
import java.util.*

class PolicyDeleteCommand(di: DI) : CliktCommand(
    name = "delete",
    help = "Delete a Policy"
), DIAware {

    override val di by findOrSetObject { di }

    private val service: PolicyService by instance()

    private val id: UUID by argument()
        .convert { toUuid(it) }

    override fun run() {
        val delete = service.findById(id)
        service.delete(delete)
    }
}